repropagate changes from propose-api-ref#69 (#56)
Repropagate changes to object-storage-service-api-ref from PreProd environment

This is a manually created Pull Request for changes to object-storage-service-api-ref to promote changes down the pipeline to Production environment.
This PR is created in place of #55 because of the patch conflicts.
The original change can be found under https://gitea.eco.tsi-dev.otc-service.com/docs-swiss/object-storage-service/pulls/69.

Reviewed-by: marmulle2