add changes from propose-api-ref#68 (#61)
Repropagate changes to object-storage-service-api-ref from PreProd environment

This is a manually created Pull Request for changes to object-storage-service-api-ref to promote changes down the pipeline to Production environment.
The original change can be found under https://gitea.eco.tsi-dev.otc-service.com/docs-swiss/object-storage-service/pulls/68.
This PR was created in place of #58.

Reviewed-by: marmulle2